Warning destruction.

+++ eugene/src/Prediction.cc
@@ -76,7 +76,7 @@ void Prediction :: plotPred ()
   index = 0;
 
   for (int i=0; i<SeqLen; i++)
-    PlotBarI(i, State2Phase[getNextState(i)], 0.4, PredWidth,
+    PlotBarI(i, State2Phase[(int)getNextState(i)], 0.4, PredWidth,
 	     1+((getNextState(i)==12) || (getNextState(i)==17))*3);
 }
